+++ crates/scheduler/src/call/store/fragment.rs
@@ -39,7 +39,7 @@ pub fn process<B: Board>(call: Api<DispatchSchedulerCall<B>>) {
 fn insert<B: Board>(mut call: SchedulerCall<B, api::insert::Sig>) {
     let api::insert::Params { keys, ptr, len } = call.read();
     let scheduler = call.scheduler();
-    let memory = scheduler.applet.memory();
+    let memory = scheduler.applet.as_mut().unwrap().memory();
     let result = try {
         let keys = decode_keys(keys)?;
         let value = memory.get(*ptr, *len)?;
@@ -61,7 +61,7 @@ fn remove<B: Board>(mut call: SchedulerCall<B, api::remove::Sig>) {
 fn find<B: Board>(mut call: SchedulerCall<B, api::find::Sig>) {
     let api::find::Params { keys, ptr: ptr_ptr, len: len_ptr } = call.read();
     let scheduler = call.scheduler();
-    let mut memory = scheduler.applet.memory();
+    let mut memory = scheduler.applet.as_mut().unwrap().memory();
     let result = try {
         match fragment::read(&scheduler.store, &decode_keys(keys)?).map_err(convert)? {
             None => false,